+++ src/jtl/Connector/Shopware/Controller/CustomerOrder.php
@@ -42,6 +42,7 @@ class CustomerOrder extends DataController
             $result = array();
             $limit = $queryFilter->isLimit() ? $queryFilter->getLimit() : 100;
 
+            $shopMapper = Mmc::getMapper('Shop');
             $mapper = Mmc::getMapper('CustomerOrder');
             $productMapper = Mmc::getMapper('Product');
             $orders = $mapper->findAll($limit);
@@ -70,20 +71,23 @@ class CustomerOrder extends DataController
                     }
 
                     // Locale
-                    $localeSW = LocaleUtil::get((int)$orderSW['languageIso']);
-                    if ($localeSW !== null) {
-                        //$order->setLocaleName($localeSW->getLocale());
-                        $order->setLanguageISO(LanguageUtil::map($localeSW->getLocale()));
+                    $shop = $shopMapper->find((int) $orderSW['languageIso']);
+                    //$localeSW = LocaleUtil::get((int) $orderSW['languageIso']);
+                    //if ($localeSW !== null) {
+                    if ($shop !== null) {
+                        //$order->setLanguageISO(LanguageUtil::map($localeSW->getLocale()));
+                        $order->setLanguageISO(LanguageUtil::map($shop->getLocale()->getLocale()));
                     }
 
                     foreach ($orderSW['details'] as $detailSW) {
-                        $orderItem = Mmc::getModel('CustomerOrderItem');
-                        $orderItem->map(true, DataConverter::toObject($detailSW, true));
 
                         // Tax Free
                         if ((int) $orderSW['taxFree'] == 1) {
-                            $orderItem->setVat(0.0);
+                            $detailSW['taxRate'] = 0.0;
                         }
+
+                        $orderItem = Mmc::getModel('CustomerOrderItem');
+                        $orderItem->map(true, DataConverter::toObject($detailSW, true));
 
                         $detail = $productMapper->findDetailBy(array('number' => $detailSW['articleNumber']));
                         if ($detail !== null) {